OR-5592 Clean-Up(s)
