+++ CombineDemo/CombineDemo/Extensions/Reusable+Identifier.swift
@@ -1,0 +1,20 @@
+//
+//  Reusable+Identifier.swift
+//  CombineDemo
+//
+//  Created by Manish Rathi on 19/08/2023.
+//
+
+import UIKit
+
+protocol Reusable {
+    static var reuseIdentifier: String { get }
+}
+
+extension Reusable {
+    static var reuseIdentifier: String {
+        return String(describing: self)
+    }
+}
+
+extension UITableViewCell: Reusable {}
