+++ CombineDemo/CombineDemo/Extensions/UITableView+DeQueue.swift
@@ -1,0 +1,18 @@
+//
+//  UITableView+DeQueue.swift
+//  CombineDemo
+//
+//  Created by Manish Rathi on 19/08/2023.
+//
+
+import UIKit
+
+extension UITableView {
+    func register<T: UITableViewCell>(_ name: T.Type) {
+        register(T.self, forCellReuseIdentifier: T.reuseIdentifier)
+    }
+
+    func dequeueReusableCell<T: UITableViewCell>(for indexPath: IndexPath) -> T {
+        dequeueReusableCell(withIdentifier: T.reuseIdentifier, for: indexPath) as! T
+    }
+}
