+++ CombineDemo/CombineDemo/Modules/PokemonList/PokemonTableViewCell.swift
@@ -82,33 +82,7 @@ class PokemonTableViewCell: UITableViewCell {
 
         imageLoader
             .imageLoadingPublisher
-            .sink { [weak self] in self?.pokemonImageView.image = $0 }
+            .assign(to: \.image, on: pokemonImageView)
             .store(in: &cancellables)
-    }
-}
-
-protocol Reusable {
-    static var reuseIdentifier: String { get }
-}
-
-extension Reusable {
-    static var reuseIdentifier: String {
-        return String(describing: self)
-    }
-}
-
-extension PokemonTableViewCell: Reusable {}
-
-extension Pokemon.ApiResponse.Result {
-    private enum Constants {
-        static let imageURLPath = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/"
-    }
-
-    var imageURL: URL? {
-        guard let pokemonID = Int(url.lastPathComponent) else { return nil }
-
-        let imageURLString = Constants.imageURLPath + "\(pokemonID).png"
-
-        return URL(string: imageURLString)
     }
 }
